Cancel dialog to reset dialog internal model
